rename namespace Timer to Loop
