+++ src/Lib/Loop/LoopProcess/KeyboardCancelLoop.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace PhpProfiler\Lib\Timer\LoopProcess;
+namespace PhpProfiler\Lib\Loop\LoopProcess;
 
-use PhpProfiler\Lib\Timer\LoopProcessInterface;
+use PhpProfiler\Lib\Loop\LoopProcessInterface;
 
 final class KeyboardCancelLoop implements LoopProcessInterface
 {
